+++ src/Phan/Language/Type/GenericArrayInterface.php
@@ -1,7 +1,6 @@
 <?php declare(strict_types=1);
 namespace Phan\Language\Type;
 
-use Phan\Language\Type;
 use Phan\Language\UnionType;
 
 /**
@@ -10,6 +9,4 @@ use Phan\Language\UnionType;
 interface GenericArrayInterface
 {
     public function genericArrayElementUnionType() : UnionType;
-
-    public function genericArrayElementType() : Type;
 }